isolated key_gen(rand), changed make_positive and R(), added Doc comments
